Backend/requirements.txt
# Core Framework (Gỡ ghim phiên bản để giải quyết xung đột)
fastapi
uvicorn[standard]

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
chromadb
numpy<2.0
alembic==1.13.1

# Configuration
python-dotenv==1.0.0

# Authentication & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
bcrypt==4.1.1

# Data Validation
pydantic-settings

# Google APIs
google-generativeai==0.8.5
google-api-python-client==2.110.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.12.4

# OpenAI
openai==1.40.0

# Data Processing
Pillow==10.1.0

# File Processing
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2
beautifulsoup4==4.12.2
docx2txt==0.9
pandas

# Langchain
langchain-text-splitters


# Networking
requests
websockets==12.0

# Async File Operations
aiofiles==23.2.1

# Cache
redis==5.0.1